requirements.txt
fastapi
uvicorn[standard]
pydantic
email-validator
python-multipart
python-dotenv
